pyproject.toml
[tool.poetry]
name = "audit-to-csv"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
pysnyk = "^0.6.3"
typer = "^0.4.0"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
